+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Added
+- Added `craft\helpers\Db::validateDatabaseTimezoneSupport()`.
+
 ### Fixed
 - Fixed a bug where eager-loading entrys’ authors wasn’t including pending and suspended authors. ([#7963](https://github.com/craftcms/cms/issues/7963))